(#2939) - refactor checkpointer to own file
